# Conflicts:
#	Assets/00.Scenes/Unit.unity.meta
#	Assets/01.Scripts/Test.meta
#	Assets/01.Scripts/Unit.meta
#	Assets/03.Prefabs/Player.prefab.meta
#	Assets/99.Assets/GabrielAguiarProductions/VFXGraph_Hits_2020.3_HDRP_v1.unitypackage.meta
#	Assets/99.Assets/GabrielAguiarProductions/VFXGraph_Hits_2020.3_URP_v1.unitypackage.meta

+++ Assets/01.Scripts/Enemy/Attack/IDamageable.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    public void Damaged(int damage);
+    public void Damaged(float damage);
 }
